+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Version 1.0.5 (10/26/2023)
+
+*	[edcre-v1.0.5-windows\_x86](https://github.com/alex-free/edcre/releases/download/v1.0.5/edcre-v1.0.5-windows-x86.zip) _For Windows 95 OSR 2.5 Or Newer (32-bit Windows)_
+*	[edcre-v1.0.5-windows-x86\_64](https://github.com/alex-free/edcre/releases/download/v1.0.5/edcre-v1.0.5-windows-x86_64.zip) _For 64-bit Windows_
+*	[edcre-v1.0.5-linux-x86](https://github.com/alex-free/edcre/releases/download/v1.0.5/edcre-v1.0.5-linux-x86_static.zip) _For x86 Linux Distros_
+*	[edcre-v1.0.5-linux-x86\_64](https://github.com/alex-free/edcre/releases/download/v1.0.5/edcre-v1.0.5-linux-x86_64_static.zip) _For x86_64 Linux Distros_
+*	[edcre-v1.0.5-source](https://github.com/alex-free/edcre/archive/refs/tags/v1.0.5.zip)
+
+Changes:
+
+*  Removed the `-z` argument. EDCRE now starts EDC/ECC regeneration at sector zero by default, this can be modified for any sector number starting by using the new `-s` argument followed by a sector number, i.e. `-s 16`.
+
+*   Significantly improved argument handling and cleaned up code. 
+
 ## Version 1.0.4 (9/7/2023)
 
 *	[edcre-1.0.4-windows\_x86](https://github.com/alex-free/edcre/releases/download/v1.0.4/edcre-1.0.4-windows_x86.zip) _For Windows 95 OSR 2.5 Or Newer (32-bit Windows)_